# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EdgeX Ops Intelligence - Liquidity API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/activity/events": {
      "get": {
        "parameters": [
          {"in": "query", "name": "platform", "schema": {"type": "string"}},
          {"in": "query", "name": "activity_type", "schema": {"type": "string"}},
          {"in": "query", "name": "status", "schema": {"type": "string"}},
          {"in": "query", "name": "review_status", "schema": {"type": "string"}},
          {"in": "query", "name": "limit", "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "activity events",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivityEventListResponse"}}}
          },
          "503": {
            "description": "activity disabled",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivityUnavailableResponse"}}}
          }
        }
      }
    },
...
    "/api/listing/candidates": {
      "get": {
        "parameters": [
          {"in": "query", "name": "limit", "schema": {"type": "integer"}},
          {"in": "query", "name": "status", "schema": {"type": "string"}},
          {"in": "query", "name": "evidence_kind", "schema": {"type": "string"}},
          {"in": "query", "name": "platform", "schema": {"type": "string"}},
          {"in": "query", "name": "symbol", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "listing candidates",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListingCandidateListResponse"}}}
          },
          "503": {
            "description": "listing disabled",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListingUnavailableResponse"}}}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityEventSummary": {
        "type": "object",
        "additionalProperties": true,
        "required": ["id", "platform", "source_group", "title", "activity_type", "review_status", "event_version", "content_hash", "dedupe_key"],
        "properties": {
          "id": {"type": "integer"},
          "platform": {"type": "string"},
          "source_group": {"type": "string"},
          "source_url": {"type": "string", "nullable": true},
          "title": {"type": "string"},
          "activity_type": {"type": "string"},
          "review_status": {"type": "string"},
          "ops_decision_action": {"type": "string", "nullable": true},
          "event_status": {"type": "string"},
          "event_version": {"type": "integer"},
          "content_hash": {"type": "string"},
          "dedupe_key": {"type": "string"},
          "publish_time": {"type": "string", "format": "date-time", "nullable": true},
          "needs_human_review": {"type": "boolean"},
          "auto_push_allowed": {"type": "boolean"}
        }
      },
...
      "ActivityEventListResponse": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/ActivityEventSummary"}},
          "next_cursor": {"type": "string"}
        }
      },
...
      "ActivityUnavailableResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "feature": {"type": "string"},
          "reason": {"type": "string"}
        }
      },
      "ListingCandidate": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "id": {"type": "integer"},
          "canonical_symbol": {"type": "string"},
          "display_symbol": {"type": "string"},
          "market_surface": {"type": "string"},
          "instrument_kind": {"type": "string"},
          "lifecycle_status": {"type": "string"},
          "lifecycle_status_label": {"type": "string"},
          "evidence_kind": {"type": "string"},
          "confidence_level": {"type": "string"},
          "business_score": {"type": "number", "nullable": true},
          "business_score_version": {"type": "string"},
          "recommendation": {"type": "string"},
          "recommendation_label": {"type": "string"},
          "source_platforms": {"type": "array", "items": {"type": "string"}},
          "first_observed_at": {"type": "string", "format": "date-time"},
          "last_observed_at": {"type": "string", "format": "date-time"}
        }
      },
      "ListingCandidateListResponse": {
        "type": "object",
        "required": ["candidates", "count"],
        "properties": {
          "candidates": {"type": "array", "items": {"$ref": "#/components/schemas/ListingCandidate"}},
          "count": {"type": "integer"}
        }
      },
...
      "ListingUnavailableResponse": {
        "type": "object",
        "required": ["status", "feature"],
        "properties": {
          "status": {"type": "string"},
          "reason": {"type": "string"},
          "feature": {"type": "string"}
        }
      }
    }
  }
}