# web/lib/api/types.ts
import type { components } from './types.gen';

type Schema<Name extends keyof components['schemas']> = components['schemas'][Name];

export type ApiStatus = Schema<'ApiStatus'>;
export type DataFreshness = Schema<'DataFreshness'>;
export type AssetCategoryKey = string;
export type SymbolMapping = Schema<'SymbolMapping'>;
export type DashboardCategorySymbol = Schema<'DashboardCategorySymbol'>;
export type DashboardCategory = Schema<'DashboardCategory'>;
export type SymbolsResponse = Schema<'SymbolsResponse'>;
export type PlatformRow = Schema<'PlatformRow'>;
export type PlatformFundingRate = Schema<'PlatformFundingRate'>;
export type DepthTierMetrics = Schema<'DepthTierMetrics'>;
export type CoinGeckoLineage = Schema<'CoinGeckoLineage'>;
export type DataSources = Schema<'DataSources'>;
export type OpsIntelligenceMeta = Schema<'OpsIntelligenceMeta'>;
export type LiquidityKPIs = Schema<'LiquidityKPIs'>;
export type LiquiditySnapshot = Schema<'LiquiditySnapshot'>;
export type QualityKPIs = Schema<'QualityKPIs'>;
export type QualitySnapshot = Schema<'QualitySnapshot'>;
export type ShareRow = Schema<'ShareRow'>;
export type ShareSnapshot = Schema<'ShareSnapshot'>;
export type Top30Row = Schema<'Top30Row'>;
export type Top30Snapshot = Schema<'Top30Snapshot'>;
export type Top30AggregateRow = Schema<'Top30AggregateRow'>;
export type Top30DivergenceRow = Schema<'Top30DivergenceRow'>;
export type Top30DivergenceKPI = Schema<'Top30DivergenceKPI'>;
export type Top30DivergenceSnapshot = Schema<'Top30DivergenceSnapshot'>;
export type ActivityEventSummary = Schema<'ActivityEventSummary'>;
export type ActivityEventSymbol = Schema<'ActivityEventSymbol'>;
export type ActivityEventListResponse = Schema<'ActivityEventListResponse'>;
export type ActivityEventDetailResponse = Schema<'ActivityEventDetailResponse'>;
export type ActivitySourceHealth = Schema<'ActivitySourceHealth'>;
export type ActivitySourceHealthResponse = Schema<'ActivitySourceHealthResponse'>;
export type ActivityDelivery = Schema<'ActivityDelivery'>;
export type ActivityDeliveryListResponse = Schema<'ActivityDeliveryListResponse'>;
export type ActivityDecisionRequest = Schema<'ActivityDecisionRequest'>;
export type ActivityMutationResponse = Schema<'ActivityMutationResponse'>;